(SERVER-2265) Refactor: moved clean
