+++ lib/puppetserver/ca/action/clean.rb
